requirements.txt
anyio==3.7.0
APScheduler==3.10.1
apturl==0.5.2
arabic-reshaper==3.0.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.7.2
asn1crypto==1.5.1
asttokens==2.2.1
attrs==23.1.0
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.12.2
black==23.3.0
bleach==6.0.0
blinker==1.4
boto3==1.26.114
botocore==1.29.114
Brlapi==0.8.3
certifi==2023.5.7
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.1.0
chrome-gnome-shell==0.0.0
click==8.1.3
colorama==0.4.4
comm==0.1.3
command-not-found==0.3
contourpy==1.1.0
crispy-bootstrap5==0.7
cryptography==40.0.1
cssselect2==0.7.0
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.2.18
debugpy==1.6.7
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.4
distro==1.7.0
distro-info===1.1build1
Django==4.2.2
django-cors-headers==3.14.0
django-crispy-forms==2.0
django-filter==23.2
django-scheduler==0.10.1
django-social-auth==0.7.28
django-storages==1.13.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.3
duplicity==0.8.21
exceptiongroup==1.1.1
executing==1.2.0
fasteners==0.14.1
fastjsonschema==2.17.1
filelock==3.6.0
fonttools==4.40.0
fqdn==1.5.1
future==0.18.2
GDAL==3.4.1
gyp==0.1
hijra==1.0
html5lib==1.1
httplib2==0.20.2
icalendar==5.0.7
idna==3.4
importlib-metadata==4.6.4
inflection==0.5.1
ipykernel==6.23.3
ipython==8.14.0
ipython-genutils==0.2.0
ipywidgets==8.0.6
isoduration==20.11.0
jedi==0.18.2
jeepney==0.7.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jsonpointer==2.4
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyter_server==2.6.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
keyring==23.5.0
kiwisolver==1.4.4
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
lxml==4.9.2
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
matplotlib==3.7.1
matplotlib-inline==0.1.6
mistune==3.0.1
monotonic==1.6
more-itertools==8.10.0
mypy-extensions==1.0.0
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.6.0
nbformat==5.9.0
nest-asyncio==1.5.6
netifaces==0.11.0
notebook==6.5.4
notebook_shim==0.2.3
numpy==1.21.5
oauth2==1.9.0.post1
oauthlib==3.2.0
olefile==0.46
oscrypto==1.3.0
overrides==7.3.1
packaging==23.0
pandas==2.0.2
pandocfilters==1.5.0
paramiko==2.9.3
parso==0.8.3
pathspec==0.11.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==2.6.2
prometheus-client==0.17.0
prompt-toolkit==3.0.38
protobuf==3.12.4
psutil==5.9.5
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pycairo==1.20.1
pycparser==2.21
pycups==2.0.1
Pygments==2.15.1
PyGObject==3.42.1
pyHanko==0.18.1
pyhanko-certvalidator==0.22.0
PyJWT==2.6.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pypdf==3.8.1
pypng==0.20220715.0
pyRFC3339==1.1
pyrsistent==0.19.3
python-apt==2.4.0+ubuntu1
python-bidi==0.4.2
python-dateutil==2.8.2
python-debian===0.1.43ubuntu1
python-dotenv==1.0.0
python-json-logger==2.0.7
python-openid==2.2.5
python-social-auth==0.3.6
python3-openid==3.2.0
pytz==2023.3
pyxdg==0.27
PyYAML==6.0
pyzmq==25.1.0
qrcode==7.4.2
qtconsole==5.4.3
QtPy==2.3.1
reportlab==3.6.8
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
s3transfer==0.6.0
scikit-learn==1.2.2
scipy==1.10.1
SecretStorage==3.3.1
Send2Trash==1.8.2
six==1.16.0
sniffio==1.3.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
soupsieve==2.4.1
sqlparse==0.4.4
stack-data==0.6.2
svglib==1.5.1
systemd-python==234
terminado==0.17.1
threadpoolctl==3.1.0
tinycss2==1.2.1
tomli==2.0.1
tornado==6.3.2
traitlets==5.9.0
typing_extensions==4.6.3
tzdata==2023.3
tzlocal==5.0.1
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
uri-template==1.3.0
uritemplate==4.1.1
uritools==4.0.1
urllib3==2.0.3
usb-creator==0.3.7
virtualenv==20.13.0+ds
wadllib==1.3.6
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.1
widgetsnbextension==4.0.7
xdg==5
xhtml2pdf==0.2.11
xkit==0.0.0
zipp==1.0.0
